Drop pulldown-cmark
