+++ Cargo.toml
@@ -13,7 +13,6 @@ html-escape = "0.2"
 json5 = "0.3"
 lazy_static = "1.4"
 lx-json-feed = { path = "./crates/json-feed" }
-pulldown-cmark = { version = "0.8", default-features = false }
 rayon = "1.5.0"
 regex = "1.4"
 serde = "1.0"
